refactor: move addTaxesToGroup func to its own file
