+++ imports/plugins/core/orders/server/no-meteor/mutations/placeOrder.js
@@ -9,6 +9,7 @@ import { Address as AddressSchema, Order as OrderSchema, Payment as PaymentSchem
 import getDiscountsTotalForCart from "/imports/plugins/core/discounts/server/no-meteor/util/getDiscountsTotalForCart";
 import xformOrderGroupToCommonOrder from "/imports/plugins/core/orders/server/util/xformOrderGroupToCommonOrder";
 import { getPaymentMethodConfigByName } from "/imports/plugins/core/payments/server/no-meteor/registration";
+import addTaxesToGroup from "../../util/addTaxesToGroup";
 import verifyPaymentsMatchOrderTotal from "../../util/verifyPaymentsMatchOrderTotal";
 
 const orderItemsSchema = new SimpleSchema({
@@ -292,45 +293,6 @@ async function addShipmentMethodToGroup(context, finalGroup, cleanedInput, group
     handling: selectedFulfillmentMethod.handlingPrice,
     rate: selectedFulfillmentMethod.shippingPrice
   };
-}
-
-/**
- * @summary Adds taxes to the final fulfillment group
- * @param {Object} context - an object containing the per-request state
- * @param {Object} finalGroup Fulfillment group object pre shipment method addition
- * @param {Object} cleanedInput - Necessary orderInput. See SimpleSchema
- * @param {Object} groupInput - Original fulfillment group that we compose finalGroup from. See SimpleSchema
- * @param {String} discountTotal - Calculated discount total
- * @param {String} orderId - Randomized new orderId
- * @returns {Object} Fulfillment group object post tax addition
- */
-async function addTaxesToGroup(context, finalGroup, cleanedInput, groupInput, discountTotal, orderId) {
-  const { collections } = context;
-  const { order: orderInput } = cleanedInput;
-  const { billingAddress, cartId, currencyCode } = orderInput;
-
-  const commonOrder = await xformOrderGroupToCommonOrder({
-    billingAddress,
-    cartId,
-    collections,
-    currencyCode,
-    group: finalGroup,
-    orderId,
-    discountTotal
-  });
-
-  const { itemTaxes, taxSummary } = await context.mutations.getFulfillmentGroupTaxes(context, { order: commonOrder, forceZeroes: true });
-  finalGroup.items = finalGroup.items.map((item) => {
-    const itemTax = itemTaxes.find((entry) => entry.itemId === item._id) || {};
-
-    return {
-      ...item,
-      tax: itemTax.tax,
-      taxableAmount: itemTax.taxableAmount,
-      taxes: itemTax.taxes
-    };
-  });
-  finalGroup.taxSummary = taxSummary;
 }
 
 /**
